requirements.txt
fastapi
uvicorn
google-auth-oauthlib
google-api-python-client
python-dotenv
langchain-groq
langchain-ollama
langchain-core
langchain
langgraph
